Add support for Colloquy-style nick autocompletion without entering any string.
In Colloquy, nick autocompletion can be started without entering any characters from the target user's
nick name (i.e., but just pressing Tab at the start of the line).  This is useful when you're having a
conversation with somebody in a crowded channel with lots of people talking: you can just hit tab before
starting your line, and the client will mention the nick of the correct user for you.

This patch implements the same idea in LimeChat.  The implementation strategy is simple.  While
generating the array of candidate nicks, we just look at the first user's weight, and if it's greater
than 0, we will artificially populate the array with only that nick, which enables us to use the same
logic for the rest of the nick autocompletion function untouched.

With this patch, the quality of the auto-completion is much better than Colloquy, because LimeChat has
the notion of conversations based on user's weight.  So random people pinging you when you're in the
middle of a conversation wouldn't distract you.