allow user to ignore inline images from certain nicks
This is helpful when you have a specific person who consistently posts
images that are not pleasing to you.

A text field has been added to the Preferences where the user
types in a comma-separated list of nicks to ignore. I wanted to
make this a table like the highlight words, but it didn't fit
into the space on the Log tab. The plumbing is set up to switch
it out for a table in the future, if someone desires to do it.