Thumbnail support for regular pixiv image URLs
Regular pixiv URLs don't have enough information in the URL to derive the
image's URL, so we have to do some HTML parsing. Since we can't block or the UI
freezes, we do the request as part of ImageSizeCheckClient's check.

Pixiv URLs also need a bit of special treatment. Clicking the thumbnail would
open the direct link to the image in the browser, which isn't very useful because
you would get a 403 back (or a small image), so we inject a link to the web
interface for the image.