Add sanity check for _response in ImageSizeCheckClient.m
When Limechat loads the contents of a URL it believes to be an image,
it follows redirects (the “Location” HTTP header). If a URL redirects
to a data image, such as:

data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAA
IBRAA7

then the value of the _response variable in the ImageSizeCheckClient.m
source file is NOT an instance of NSHTTPURLResponse. As a result,
calling -allHeaderFields or -statusCode on _response results in a crash.