ensure authzid and authcid are equal in SASL PLAIN
https://datatracker.ietf.org/doc/html/rfc4616#section-2
https://datatracker.ietf.org/doc/html/rfc4422#section-2

authzid ("authorization id") is optional, authcid ("authentication id") is
mandatory; best practice in IRC is to omit authzid or set it equal to
authcid. Previous Limechat behavior was to send the nickname as the authzid
which will break authentication on major networks (e.g. Libera) if the
nickname is not registered with NickServ.